Adding an "Adventure Demo" level and docs (#3)
The Adventure Demo provides design patterns and an example of how to use Ink and Inkpot to manage narrative text for a Narrative Adventure game in Unreal Engine.

It covers design patterns for implementing:
- Dialogue
- Player Journal
- Quests

### AdventureDemo.ink and .uasset
The ink file for the Adventure Demo and Inkpot Story asset

### AdventureDemo.umap 
The actual Adventure Demo level.

There's some basic path-switching logic in the level BP.

### WBP_Display_Adventure

An expanded version of WBP_Display that includes:
- The Journal panel, Journal button, and JournalUpdated notification
- Logic for opening/closing the journal
- Logic for showing Journal Updates, using a Story Change Delegate
- Visual Updates to the main story box, and logic to hide it when there's no narrative text to show
- Logic to customize the timer for each line using a `linetime: ` tag in Ink

A lot of this code is duplicated from Inkpot Demo. This is inelegant, but @TCR-Nick is about to overhaul the demo level, so it's easier to keep the Adventure Demo separate until his changes are done.

### BP_AdventureDemoGameMode:

This is a copy of `InkpotDemoGameMode` that points to `WBP_Display_Adventure` in place of `WBP_Display` widget blueprint.

Again, inelegant--but it's better to wait until the main Demo level is updated before consolidating.

### Updates to README and the Intro level
Added AdventureDemo to the list of levels in the Intro level and main repo Readme.

### README-AdventureDemo.md

Honestly this might be a bit much. Documentation escaped containment.

### AD_* images in Documentation/

Screenshots for README-AdventureDemo.md